began implementing api integration and update dropdowns a bit
