+++ src/components/Navigation.scss
@@ -34,7 +34,7 @@ text-decoration: none;
 
 	&.is-active,
 	&:hover {
-		color: rgb(20,41,83);
+		color: #0000ee;
 		;
 	}
