+++ src/pages/Teams.scss
@@ -1,0 +1,16 @@
+// CSS Reset
+@import "~reset-css/reset.css";
+@import "~semantic-ui-css/semantic.css";
+@import url(https://fonts.googleapis.com/css?family=Montserrat);
+
+// Box sizing
+*, *:before, *:after {
+	box-sizing: inherit;
+}
+body {
+	box-sizing: border-box;
+}
+
+.team-submit-form {
+	width: 100%;
+}
